Moved test out into a ticket test
